# swagger.yaml
openapi: 3.0.3
info:
  title: Eco Tips - OpenAPI 3.0
  description: |-
    Eco Tips is a web application that allows its users, in a fun way, to reduce their ecological impact and save money.

    In the current context, more and more citizens are aware of the climate issues and need to take action in their daily lives to reduce their carbon footprint. The country's economic context also leads citizens to reduce their daily expenses. To meet this need, Eco Tips offers its users eco-gestures based in part on a [study by CLER - Network for Energy Transition](https://cler.org/wp-content/uploads/2021/01/Guide-100-e%CC%81cogestes-De%CC%81clics_planches.pdf), helping them to save money and live more sustainably.

    Therefore, Eco Tips provides a fun solution that encourages individuals to awaken their environmental awareness, to sustainably install eco-gestures in their daily lives in order to reduce the users' carbon footprint. The application informs users about new uses while creating a community of committed citizens.
    
    Some useful links:
    - [Eco Tips back Github repository](https://github.com/O-clock-Ohm/projet-03-eco-tips-back)
    - [Eco Tips front Github repository](https://github.com/O-clock-Ohm/projet-03-eco-tips-front)
servers:
  - url : http://tonympt-server.eddi.cloud:8080/
...
paths:
...
  /me/achievement:
    post:
      tags:
        - achievement
      summary: Adds an achievement
      description: Adds an achievement
      operationId: addAchievement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Achievement'
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /achievement/proposal:
    get:
      tags:
        - achievement
      summary: Returns all submitted achievements (admin only)
      description: Returns all submitted achievements
      operationId: getAllProposalAchievement
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /achievement/proposal/{id}:
    patch:
      tags:
        - achievement
      summary: Updates an achievement to an approved state (admin only)
      description: Updates an achievement to an approved state
      operationId: updateProposalAchievementToFalse
      responses:
        '204':
          description: Successful operation
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /achievement:
    get:
      tags:
        - achievement
      summary: Returns all approved achievements (admin only)
      description: Returns all approved achievements
      operationId: getAllNotProposalAchievement
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /achievement/{id}:
    patch:
      tags:
        - achievement
      summary: Updates an achievement (admin only)
      description: Updates an achievement
      operationId: updateAchievement
      responses:
        '204':
          description: Successful operation
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
    delete:
      tags:
        - achievement
      summary: Deletes an achievement (admin only)
      description: Deletes an achievement 
      operationId: deleteAchievement
      responses:
        '204':
          description: Successful operation
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /achievement/random:
    get:
      tags:
        - achievement
      summary: Returns a random approved achievement
      description: Returns a random approved achievement
      operationId: getOneRandomAchievement
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
components:
  schemas:
...
    Achievement:
      type: object
      properties:
        title:
          type: string
        image:
          type: string
          example: data:image/jpeg;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
        description:
          type: string